Added SFRAT dataframe conversions to dataClass.py
For use with trend tests. Trend tests match SFRAT for SYS1 data. Renamed 'Cumulative' column in imported dataframe to 'CFC'. Added variable in dataClass for the number of columns that will always be present, used for determining which columns are covariates. Currently, there are 5 static columns: T, FC, CFC, FN, IF.